add test hash class.

+++ src/utils/hash/hash_fn.h
@@ -1,10 +1,10 @@
 #ifndef mocha_hash_fn_h_
 #define mocha_hash_fn_h_
-
+#include <utils/int_types.h>
 namespace mocha {
 
-static inline uint32_t GetHash( uint32_t key ) {
-  uint32_t hash = key;
+static inline uint64_t GetHash( uint64_t key ) {
+  uint64_t hash = key;
   hash = ~hash + (hash << 15);  // hash = (hash << 15) - hash - 1;
   hash = hash ^ (hash >> 12);
   hash = hash + (hash << 2);
